chore: Extract colors to lib
