+++ nextjs/src/lib/colors.ts
@@ -1,0 +1,14 @@
+export const colors = {
+  stone: "#0F0F0F",
+  biscay: "#1C2E5D",
+  sapphire: "#2E4B98",
+  jumbo: "#8B8B8C",
+  neutral: "#F5F6F5",
+  sunglow: "#FFCA31",
+  sky: "#55C0EE",
+  cinnamon: "#E56820",
+  green: "#34AE6B",
+  razzmatazz: "#E30D74",
+  fuchsia: "#7E57C2",
+  manhattan: "#F7C6A5",
+} as const
